Extract stream to its own module.
